migrated all da meta
