--- app/DefaultTags.tsx
@@ -1,18 +1,0 @@
-const DefaultTags = () => {
-  return (
-    <>
-      <meta name="viewport" content="width=device-width, initial-scale=1" />
-      <link href="/favicon.ico" rel="shortcut icon" />
-      <meta property="og:url" content={process.env.NEXT_PUBLIC_SERVER} />
-      <meta property="og:site_name" content="Haruhi" />
-      <meta property="og:type" content="website" />
-      <meta
-        name="keywords"
-        content="Haruhi, Anime, Manga, Streaming, Free, NoAds, Ads, Anilist, Flamindemiod, React, NextJS, Online"
-      />
-      <meta name="robots" content="index, follow" />
-    </>
-  );
-};
-
-export default DefaultTags;
